Fixed User addition to lead to registration flow plus minor cleaning up of code
